\name{dist_get}
\alias{dist_get}
\title{Retrieve distances from a `dist` object.}
\usage{
  dist_get(d, idx1, idx2)
}
\arguments{
  \item{d}{An object of class `dist`.}

  \item{idx1,idx2}{Indicies specifying the distances to
  extract.}
}
\description{
  Retrieve distances from a `dist` object.
}

