\name{dist_subset}
\alias{dist_subset}
\title{Extract parts of a `dist` object.}
\usage{
  dist_subset(d, idx)
}
\arguments{
  \item{d}{An object of class `dist`.}

  \item{idx}{Indices specifying the subset of distances to
  extract.}
}
\value{
  An object of class `dist`
}
\description{
  Extract parts of a `dist` object.
}

