\name{group_samples_by_column}
\alias{group_samples_by_column}
\title{Group samples by common value in metadata column.}
\usage{
  group_samples_by_column(sample.mapping, colname)
}
\arguments{
  \item{sample.mapping}{Sample mapping data frame.}

  \item{colname}{Column name by which to group.}
}
\value{
  A list of sample names per category.
}
\description{
  Group samples by common value in metadata column.
}

