\name{parse_mapping_file}
\alias{parse_mapping_file}
\title{Read a QIIME sample mapping file.}
\usage{
  parse_mapping_file(filepath)
}
\arguments{
  \item{filepath}{Path to sample mapping file.  The file
  must conform to the QIIME standards, detailed at
  \url{http://qiime.org/documentation/file_formats.html}.}
}
\value{
  A data frame.
}
\description{
  Read a QIIME sample mapping file.
}

