\name{rarefaction_stats}
\alias{rarefaction_stats}
\title{Compute summary statistics for collated alpha diversity tables.}
\usage{
  rarefaction_stats(rarefaction_table)
}
\arguments{
  \item{rarefaction_table}{A collated alpha diversity data
  frame.}
}
\value{
  A data frame of summary statistics.
}
\description{
  Compute summary statistics for collated alpha diversity
  tables.
}

