% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{read_blast_table}
\alias{read_blast_table}
\title{Parse tabular output from BLAST.}
\usage{
read_blast_table(filepath)
}
\arguments{
\item{filepath}{Path to tabular BLAST output file.}
}
\value{
A data frame of BLAST results.
}
\description{
Parse tabular output from BLAST.
}

