% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{read_qiime_distmat}
\alias{read_qiime_distmat}
\title{Read a QIIME distance matrix file.}
\usage{
read_qiime_distmat(filepath)
}
\arguments{
\item{filepath}{Path to QIIME distance matrix file.}
}
\value{
A distance matrix.
}
\description{
Read a QIIME distance matrix file.
}

