% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{read_qiime_rarefaction}
\alias{read_qiime_rarefaction}
\title{Read a collated alpha diversity table from QIIME.}
\usage{
read_qiime_rarefaction(filepath)
}
\arguments{
\item{filepath}{Path to collated alpha diversity file.}
}
\value{
A data frame with columns `SampleID`, `sequences_per_sample`,
  `iteration`, and `diversity`.
}
\description{
Read a collated alpha diversity table from QIIME.
}

