\docType{data}
\name{relmbeta_alpha}
\alias{relmbeta_alpha}
\title{Shannon diversity measurements from murine gut microbiome}
\source{
  Hildebrandt et al. High-fat diet determines the
  composition of the murine gut microbiome independently of
  obesity. \emph{Gastroenterology} \strong{137}, 1716
  (2009).
}
\description{
  \code{relmbeta_alpha} is a data frame containing the
  Shannon diversity (base 2) for each sample at a level of
  10-500 sequences per sample.
}
\details{
  The \code{\link{relmbeta}} documentation provides an
  overview of the study.
}
\seealso{
  \code{\link{relmbeta}}
}
\keyword{datasets}

