% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiimer.R
\docType{data}
\name{relmbeta_dist}
\alias{relmbeta_dist}
\title{Unweighted UniFrac distances from murine gut microbiome}
\format{An object of class \code{"dist"}.}
\source{
Hildebrandt et al. High-fat diet determines the composition of the 
  murine gut microbiome independently of obesity. \emph{Gastroenterology} 
  \strong{137}, 1716 (2009).
}
\description{
\code{relmbeta_dist} is an object of class \code{"dist"}, containing
the unweighted UniFrac distances between samples.
}
\details{
The \code{\link{relmbeta}} documentation provides an overview of the study.
}
\seealso{
\code{\link{relmbeta}}
}
\keyword{datasets}

