\name{saturated_rainbow}
\alias{saturated_rainbow}
\title{Saturated rainbow palette}
\usage{
  saturated_rainbow(n, saturation_limit = 0.4)
}
\arguments{
  \item{n}{Length of the palette}

  \item{saturation_limit}{The fraction of the total palette
  length over which the rainbow extends.  Above this limit,
  the color will remain the same.}
}
\value{
  A vector of colors.
}
\description{
  Saturated rainbow palette
}

