% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{dist_subset}
\alias{dist_subset}
\title{Extract parts of a `"dist"` object.}
\usage{
dist_subset(d, idx)
}
\arguments{
\item{d}{A distance matrix object of class `"dist"`.}

\item{idx}{Indices specifying the subset of distances to extract.}
}
\value{
A distance matrix.
}
\description{
Extract parts of a `"dist"` object.
}
\examples{
data(relmbeta_dist)
dist_subset(relmbeta_dist, c("A1", "A2", "A3", "A4", "A5"))
}

