\name{rarefaction_stats}
\alias{rarefaction_stats}
\title{Compute summary statistics for collated alpha diversity tables.}
\usage{rarefaction_stats(rarefaction_table)}
\description{Compute summary statistics for collated alpha diversity tables.}
\value{A data frame of summary statistics.}
\arguments{\item{rarefaction_table}{A collated alpha diversity data frame.}}
