% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{rarefaction_stats}
\alias{rarefaction_stats}
\title{Compute summary statistics for collated alpha diversity tables.}
\usage{
rarefaction_stats(rarefaction_table)
}
\arguments{
\item{rarefaction_table}{A collated alpha diversity data frame.}
}
\value{
A data frame of summary statistics, with columns `SampleID`, 
  `sequences_per_sample`, `diversity.mean`, and `diversity.sd`.
}
\description{
Compute summary statistics for collated alpha diversity tables.
}
\examples{
data(relmbeta_alpha)
head(rarefaction_stats(relmbeta_alpha))
}

