\docType{data}
\name{relmbeta_biom}
\alias{relmbeta_biom}
\title{BIOM format object from murine gut microbiome}
\format{A \code{"biom"} object with 227 rows and 20 columns.}
\source{
  Hildebrandt et al. High-fat diet determines the
  composition of the murine gut microbiome independently of
  obesity. \emph{Gastroenterology} \strong{137}, 1716
  (2009).
}
\description{
  The \code{relmbeta_biom} object is a representation of
  the BIOM file produced by QIIME.  It was produced by
  loading the BIOM file with the function
  \code{RJSONIO::fromJSON}.
}
\details{
  The \code{\link{relmbeta}} documentation provides an
  overview of the study.
}
\seealso{
  \code{\link{relmbeta}}
}
\keyword{datasets}

